优化删除审批流操作，增加非空判断

+++ service/src/main/java/cn/thinkfree/service/approvalflow/impl/ApprovalFlowTimeoutNoticeServiceImpl.java
@@ -38,8 +38,10 @@ public class ApprovalFlowTimeoutNoticeServiceImpl implements ApprovalFlowTimeout
 
     @Override
     public void deleteByNodeNums(List<String> nodeNums) {
-        ApprovalFlowTimeoutNoticeExample example = new ApprovalFlowTimeoutNoticeExample();
-        example.createCriteria().andNodeNumIn(nodeNums);
-        timeoutNoticeMapper.deleteByExample(example);
+        if (nodeNums != null && nodeNums.size() > 0 ) {
+            ApprovalFlowTimeoutNoticeExample example = new ApprovalFlowTimeoutNoticeExample();
+            example.createCriteria().andNodeNumIn(nodeNums);
+            timeoutNoticeMapper.deleteByExample(example);
+        }
     }
 }
